Anfrage #6009: opennn6 implementiert, indem eigene Änderungen an opennn5 übernommen wurden.

+++ opennn/config.h
@@ -57,7 +57,7 @@ namespace opennn
 {
     using namespace std;
     using namespace Eigen;
-    using type = float;
+    using type = double;
 //    typedef double type;
 }
